# openapi.yaml
openapi: 3.1.0
info:
  title: NeroCD API
  version: 0.1.0
  description: |
    Public REST API for the NeroCD automation control plane.

    All public routes are versioned under `/api/v1`. List responses use a stable
    `{ "items": [...] }` envelope. Error responses use
    `{ "code": "stable_code", "error": "message" }`.

    List endpoints return a stable `{ "items": [...] }` envelope plus
    pagination metadata: `limit`, `offset`, `count`, and `total`. Clients can
    request `limit` and `offset` query parameters. Filtering is documented per
    endpoint. Mutations consume JSON request bodies and return the created
    resource or validation errors.
servers:
  - url: http://127.0.0.1:8080
    description: Local development server
security:
  - bearerAuth: []
  - browserSession: []
paths:
...
  /api/v1/sessions:
    get:
      summary: List local sessions (system administrator)
      operationId: listSessions
      responses:
        "200":
          description: Session metadata, never raw credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    post:
      summary: Create a local session
      operationId: createSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created. The token is returned once and should be stored by the client.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        "401":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    delete:
      summary: Revoke the current session
      operationId: revokeSession
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Session revoked.
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/sessions/revoke:
    post:
      summary: Revoke a local session (system administrator)
      operationId: revokeSessionByID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeSessionRequest"
      responses:
        "200":
          description: Session revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/browser-sessions:
    post:
      summary: Create a browser session
      operationId: createBrowserSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Browser session created. The session credential is set only in the HttpOnly `nerocd_session` cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrowserSessionResponse"
        "401":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    delete:
      summary: Revoke the current browser session
      operationId: revokeBrowserSession
      security:
        - browserSession: []
      responses:
        "204":
          description: Browser session revoked and cookie expired.
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/api-tokens:
    post:
      summary: Create a platform API token
      operationId: createAPIToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/APITokenInput"
      responses:
        "201":
          description: API token created. The bearer token is returned once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APITokenRegistration"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/api-tokens/revoke:
    post:
      summary: Revoke a platform API token
      operationId: revokeAPIToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeAPITokenInput"
      responses:
        "200":
          description: API token revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIToken"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: NeroCD session token
      description: CLI session and API token credential.
    browserSession:
      type: apiKey
      in: cookie
      name: nerocd_session
      description: "Browser-only local session cookie. Unsafe cookie-authenticated requests require X-NeroCD-CSRF: 1."
...
  responses:
    Error:
      description: Stable error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          enum: [unauthenticated, forbidden, not_found, bad_request, internal_error]
        error:
          type: string
...
    CreateSessionRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    CreateSessionResponse:
      type: object
      required: [session, token]
      properties:
        session:
          $ref: "#/components/schemas/Session"
        token:
          type: string
          description: Bearer token returned once at creation time.
    BrowserSessionResponse:
      type: object
      required: [session]
      properties:
        session:
          $ref: "#/components/schemas/Session"
    Session:
      type: object
      required: [id, user_id, expires_at, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        source_ip:
          type: string
        user_agent:
          type: string
        revoked_at:
          type: string
          format: date-time
          nullable: true
    SessionListResponse:
      type: object
      required: [items, limit, offset, count, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Session"
        limit:
          type: integer
        offset:
          type: integer
        count:
          type: integer
        total:
          type: integer
    RevokeSessionRequest:
      type: object
      required: [session_id]
      properties:
        session_id:
          type: string
    APITokenInput:
      type: object
      required: [name, roles]
      properties:
        name:
          type: string
        kind:
          type: string
          enum: [service_account, bootstrap]
          default: service_account
        roles:
          type: array
          items:
            type: string
            enum: [system_admin, runner_admin]
        expires_at:
          type: string
          format: date-time
    RevokeAPITokenInput:
      type: object
      required: [token_id]
      properties:
        token_id:
          type: string
    APITokenRegistration:
      type: object
      required: [api_token, token]
      properties:
        api_token:
          $ref: "#/components/schemas/APIToken"
        token:
          type: string
          description: Bearer token returned once at creation time.
    APIToken:
      type: object
      required: [id, name, kind, roles, status, created_by, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [service_account, bootstrap]
        roles:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, revoked]
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
...